Fixed issue of two admins with matching model names
This commit fixes the issue of two admins trying to create models using
Admin::addModel() using the same name. Because the SELECT statement uses
implicit INNER JOIN to join the ACL table, if the second admin is not
present in the ACL of the already existing model, a second model is
created with the same name.